Fix attack through sectors
